Fix typo in testing code.
"Compatability" should be "Compatibility"
