+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/playlist.xml
@@ -25,18 +25,18 @@
 <playlist xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../TestConfig/playlist.xsd">
 	<include>variables.mk</include>
 	<test>
-		<testCaseName>cmdLineTester_SCHelperCompatabilityTests_win</testCaseName>
+		<testCaseName>cmdLineTester_SCHelperCompatibilityTests_win</testCaseName>
 		<variations>
 			<variation>Mode110</variation>
 			<variation>Mode610</variation>	
 		</variations>
 		<command>$(MKTREE) $(REPORTDIR); \
 	$(CD) $(REPORTDIR); \
-	cp $(Q)$(TEST_RESROOT)$(D)HelperCompatabilityTests.jar$(Q) .; \
-	$(Q)$(JDK_HOME)$(D)bin$(D)jar.exe$(Q) xf HelperCompatabilityTests.jar; \
+	cp $(Q)$(TEST_RESROOT)$(D)HelperCompatibilityTests.jar$(Q) .; \
+	$(Q)$(JDK_HOME)$(D)bin$(D)jar.exe$(Q) xf HelperCompatibilityTests.jar; \
 	$(JAVA_COMMAND) $(JVM_OPTIONS) -DJAVA_EXE='$(JAVA_COMMAND) $(JVM_OPTIONS)' -DJDK_HOME=$(Q)$(JDK_HOME)$(Q) -DCPDL=$(Q)$(P)$(Q) -DPROPS_DIR=$(Q)$(TEST_RESROOT)$(D)props_win$(Q) -DSCMODE=204 \
 	-jar $(CMDLINETESTER_JAR) \
-	-config $(Q)$(TEST_RESROOT)$(D)HelperCompatabilityTests$(XMLSUFFIX).xml$(Q) \
+	-config $(Q)$(TEST_RESROOT)$(D)HelperCompatibilityTests$(XMLSUFFIX).xml$(Q) \
 	-xids all,$(JAVA_VERSION) -xlist $(Q)$(TEST_RESROOT)$(D)exclude.xml$(Q) \
 	-nonZeroExitWhenError \
 	-outputLimit 300; \
@@ -56,19 +56,19 @@
 		</subsets>
 	</test>
 	<test>
-		<testCaseName>cmdLineTester_SCHelperCompatabilityTests_unix</testCaseName>
+		<testCaseName>cmdLineTester_SCHelperCompatibilityTests_unix</testCaseName>
 		<variations>
 			<variation>Mode110</variation>
 			<variation>Mode610</variation>	
 		</variations>
 		<command>$(MKTREE) $(REPORTDIR); \
 	$(CD) $(REPORTDIR); \
-	cp $(Q)$(TEST_RESROOT)$(D)HelperCompatabilityTests.jar$(Q) .; \
-	$(Q)$(JDK_HOME)$(D)bin$(D)jar$(Q) xf HelperCompatabilityTests.jar; \
+	cp $(Q)$(TEST_RESROOT)$(D)HelperCompatibilityTests.jar$(Q) .; \
+	$(Q)$(JDK_HOME)$(D)bin$(D)jar$(Q) xf HelperCompatibilityTests.jar; \
 	$(CONVERT_TO_EBCDIC_CMD) \
 	$(JAVA_COMMAND) $(JVM_OPTIONS) -DJAVA_EXE='$(JAVA_COMMAND) $(JVM_OPTIONS)' -DJDK_HOME=$(Q)$(JDK_HOME)$(Q) -DCPDL=$(Q)$(P)$(Q) -DPROPS_DIR=$(Q)$(TEST_RESROOT)$(D)props_unix$(Q) -DSCMODE=204 \
 	-jar $(CMDLINETESTER_JAR) \
-	-config $(Q)$(TEST_RESROOT)$(D)HelperCompatabilityTests$(XMLSUFFIX).xml$(Q) \
+	-config $(Q)$(TEST_RESROOT)$(D)HelperCompatibilityTests$(XMLSUFFIX).xml$(Q) \
 	-xids all,$(JAVA_VERSION),$(JCL_VERSION) -xlist $(Q)$(TEST_RESROOT)$(D)exclude.xml$(Q) \
 	-nonZeroExitWhenError \
 	-outputLimit 300; \
