+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_unix/Test12.props
