+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_unix/Test15.props
