+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_unix/Test3.props
