+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_unix/Test4.props
