+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_unix/Test8.props
