+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test1.props
