+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test10.props
