+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test11.props
