+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test12.props
