+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test13.props
