+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test15.props
