+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test2.props
