+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test3.props
