+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test4.props
