+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test5.props
