+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test6.props
