+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test7.props
