+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test8.props
