+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/props_win/Test9.props
