+++ test/functional/cmdLineTests/shareClassTests/SCHelperCompatibilityTests/variables.mk
